[Core][Admin] Move shipment confirmation email from core to admin
