+++ src/Sylius/Bundle/AdminBundle/Resources/config/app/config.yml
@@ -1,4 +1,5 @@
 imports:
+    - { resource: "@SyliusAdminBundle/Resources/config/app/sylius/sylius_mailer.yml" }
     - { resource: "@SyliusAdminBundle/Resources/config/grids/address_log_entry.yml" }
     - { resource: "@SyliusAdminBundle/Resources/config/grids/admin_user.yml" }
     - { resource: "@SyliusAdminBundle/Resources/config/grids/channel.yml" }
